refactor and better uptime
